DjangoRest/requirements.txt
Django==1.11.6
django-tastypie==0.14.0
djangorestframework==3.7.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.2
six==1.11.0
